backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.97",
    "dj-database-url>=2.3.0",
    "dj-rest-auth[with-social]>=7.0.1",
    "django-allauth>=65.1.0",
    "django-cors-headers>=4.6.0",
    "django-storages>=1.14.4",
    "django==4.2",
    "djangorestframework-simplejwt>=5.4.0",
    "djangorestframework>=3.15.2",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "rich>=13.9.4",
]

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar>=4.3.0",
    "drf-spectacular>=0.28.0",
    "ruff>=0.9.1",
]


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py312"


[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "F",
    "I",
    "W505",
    "PT018",
    "SIM101",
    "SIM114",
    "PGH004",
    "PLE1142",
    "RUF100",
    "F404",
    "TCH",
]
ignore = [
    "PGH004",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
